requirements.txt
Django==5.0.3
lxml==5.0.1
requests>=2.28.1
bs4>=0.0.1
colorlog>=6.7.0
whitenoise>=6.4.0
gunicorn==21.2.0;platform_system!='Windows'
celery>=5.2.7
djangorestframework>=3.14.0
django-celery-results>=2.5.0
django-extensions>=3.2.1
redis==5.0.3
psycopg2-binary==2.9.9
yara-python==4.0.5
yara-scanner==2.0.3
drf_yasg==1.21.7
pyyaml==5.3.1
androguard==4.1.1
pysast@git+https://github.com/MatrixEditor/pysast.git
lief==0.14.0
google_play_scraper==1.2.6
nibarchive==1.1.0
umbrella-py==0.0.3
caterpillar@git+https://github.com/MatrixEditor/caterpillar.git